pyproject.toml
[tool.poetry]
name = "spFA"
version = "0.3.0"
description = "Probabilistic factor analysis model with covariate guided factors"
authors = ["capraz <tuemayc@hotmail.com>"]
readme = "README.md"
packages = [{include = "spfa"}]
license = "MIT"


[tool.poetry.dependencies]
python = ">=3.8.1,<3.11.6"
torch = "^1.13.1"
pyro-ppl = "<1.8.4"
pandas = ">=1.4.2"
numpy = ">=1.22.4"
click = ">=8.1.3"
scikit-learn = ">=1.1.1"
numba = ">=0.55.2"
scipy = ">=1.8.1"
muon = ">=0.1.3"
matplotlib =">=3.5.2"
Sphinx = { version = "4.2.0", optional = true }
sphinx-rtd-theme = { version = "1.0.0", optional = true }
sphinxcontrib-napoleon = { version = "0.7", optional = true }
sphinxcontrib-bibtex = "^2.5.0"
nbsphinx = { version = "0.8.9", optional = true }
jupyter = {version = "*", optional = true}
pytest = "*"


[tool.poetry.extras]
docs = ["Sphinx", "sphinx-rtd-theme", "sphinxcontrib-napoleon", "nbsphinx"]
notebook = ["jupyter"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
